Fix makefile and reorganize directories
