+++ Makefile
@@ -2,9 +2,9 @@
 lib_dir := $(CONDA_PREFIX)/lib
 include_dir := $(CONDA_PREFIX)/include
 
-compiler_and_flags := g++-9 -std=c++2a -Wall -Wextra -march=native -O3 -pedantic -Wshadow 
+compiler_and_flags := g++-9 -std=c++2a -Wall -Wextra -march=native -O3 -pedantic -Wshadow -I$(include_dir) -Wl,-rpath=$(lib_dir) -L$(lib_dir)
 
-common := -o build/nc-timestamp-mapper -I./src/ThirdParty/ -I$(include_dir) -Wl,-rpath=$(lib_dir) -L$(lib_dir) src/TimestampMapper.cpp src/ProgressBar.cpp src/DatasetDesc.cpp src/Database.cpp src/FileReaders/NCFileReader.cpp src/CLIOptions.cpp src/main.cpp  
+common := -o build/nc-timestamp-mapper -I./src/ThirdParty/ src/TimestampMapper.cpp src/Utils/ProgressBar.cpp src/DatasetDesc.cpp src/Database.cpp src/FileReaders/NCFileReader.cpp src/CLIOptions.cpp src/main.cpp
 
 libs := -lstdc++fs -lnetcdf_c++4 -lsqlite3
